contrib/upload/requirements.txt
requests[security]
pandas
python-dotenv
